Add new NewsSpecifedSpider class for fetching a fixed list of URLs rather than reading an RSS feed or sitemap
In order to allow this to work while sharing parsing code with the
existing spiders, this commit experimentally splits out the parse_page
method from spiders/uk/guardian.py into a separate class, then uses it
as a mixin both for the existing GuardianSpider class and the new
GuardianSpecifiedSpider. None of the other spiders are converted at
this point.