Add general-purpose news template, news.tmpl
We have been working for a while with several separate templates for
setting up different kinds of crawlers - one for NewsSitemapSpiders,
one for NewsCSSSpiders and so on. As the page parsing code is
normally written as a page_parse() method on the class generated
by these templates, it's not ideal for newspapers with multiple
access methods. Maybe there are (separate) needs for an archive
crawler that goes through the entire historical sitemap.xml, and
one which reads a (shorter) RSS feed for new articles? The new
NewsSpecified approach, which takes a given list of URLs, fetches
and then parses them, works perfectly alongside the other methods.

So now we have a single 'news' template. That will set up a
'FooSpiderParser' class containing the parse_page() method, and
a (commented) series of 'SomeMethodFooSpider' classes, each of
which mixes in the parser, inherits from the appropriate base
spider classes, and come as filled in as possible. So to enable
each of these should only involve uncommenting and filling in
appropriate metadata (like the address of an RSS feed).

The old templates are still available for the time being.